fix: change x type in request handling from int to double

+++ src/main/java/lab02/data/RequestData.java
@@ -1,11 +1,11 @@
 package lab02.data;
 
-public record RequestData(int[] x, double y, double r) {
+public record RequestData(double[] x, double y, double r) {
     @Override
     public String toString() {
         StringBuilder res = new StringBuilder("x = ");
         if (x != null) {
-            for (int element : x) {
+            for (double element : x) {
                 res.append(element).append(", ");
             }
         }
